requirements.txt
gspread
google-auth
pandas
plotly
streamlit>=1.30.0
altair<5.0.0